The first expect, we  can see that since refArray is passed in through a function we know that changes to the array that are made within the function will be reflected when the array is used outside of the function.
The second expect, assignedArray is simply overwritten and thus there is a new value.

The third expect is different from the first because array.slice passes the array by value which means that changes to the array are not saved. Therefore, the original value when is the same when called.