IMPORTANT: Forgot to do the push and pop / shift and unshift seperately so this commit is for both.
Popping takes the last value in the array and pops it off returning that value. Pushing adds a value to the end of the array.
Unshifting adds a value to the beginning of the array.
Shifting takes the first value in the array and shifts it off returning that value.